New parameter
You can now define the (previously hardcoded) node type to a string and describe if you're portable, mobile, etc. This value will be passed to the reflector and shown on the dashboard (where available).
The string will also get saved in a profile, when you create a new one or rewrite existing.